pyproject.toml
[tool.poetry]
name = "kynykos-ai-agent"
version = "2.0.0"
description = "KynicOS — Sistema Operativo IA para Hospitalidad y DevSecOps"
authors = ["Julian Juarez <julianjuarezmx01@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.12"

# Telegram
python-telegram-bot = "^21.0"

# LLM — cadena pública sin Anthropic
groq = "^0.11.0"
httpx = "^0.27.0"

# Config & Utils
pydantic = "^2.4.2"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
aiofiles = "^23.2.1"
tenacity = "^9.0.0"
loguru = "^0.7.0"

# WhatsApp
twilio = "^9.0.0"

# Stripe (Fase 2)
stripe = "^10.0.0"

# Extras opcionales
boto3 = "^1.34.0"
qrcode = {extras = ["pil"], version = "^7.4.0"}
Pillow = "^10.0.0"
mcp = ">=0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
